Squash output distances into a circle
When plotting the distance in the accumulator rather
than plot the distance as read from the source image
instead plot the distance as if the source image had
been mapped into a circle. This was an attempt to
prevent the artifacts that occur at 45°. It did not
work though the artifacts in my test image remained.
I think this is because the mapping to a circle affects
large distances much more than short distances and therefore
has little effect on the test image since the artifacts are
very close to the centre of the image (0 distance).