Correct artifacting
By dividing the accumulator hits in the distance dimension between
2 pixels instead of simply bucketing in a single pixel we can get rid
of most of the artifacting caused by lines at 45 and 135 degrees.
For now I'm using a bucket score of 10 and dividing that between the
two adjacent pixels. I'd like to be a bit more accurate but increasing
the score too much leads to overflows. Possibly the solution is to store
accumulator values as int32 but I'm not sure how this would affect the
performance.

As performance goes this is costly, ops per sec has gone from ~110 to
~60 but I think the removal of artifacts will really improve the
accuracy of line finding.