Extract angle calculation from hough call
Extracted common initialisation code into a separate function that
produces a struct that can be used in multiple calls to hough for the
same input and output image sizes. Calling this method for each hough
call results has no noticeable effect on performance retaining ~60
ops per second. Interestingly re-using the struct on multiple calls has
basically no effect on performance, which now that I think about it
again makes sense because it is not in a tight loop and therefore it is
being run a negligible amount of times compared to the rest of the
algorithm. Like less than this much percent 0.0000015625 !